Remove ComputationBoundary and Double semantics (#428)
## Summary
- Remove ComputationBoundary and boundary escape APIs from the
fixed-point layer.
- Replace Double-oriented math/random/reporting paths with opaque
fixed-point semantics and MetricValue for Monte Carlo outputs.
- Migrate tests and generators to typed values and BigDecimal-only
expected-value helpers.
- Keep modules/ledger unchanged.

+++ src/test/scala/com/boombustgroup/amorfati/init/InitRandomnessSpec.scala
@@ -28,5 +28,5 @@ class InitRandomnessSpec extends AnyFlatSpec with Matchers:
     first.all.map(_.seed).distinct should have size first.all.size
     first.all.map(_.seed) should not equal third.all.map(_.seed)
     first.firms.network.newStream().nextLong() shouldBe second.firms.network.newStream().nextLong()
-    first.immigration.initialStock.newStream().nextDouble() shouldBe second.immigration.initialStock.newStream().nextDouble()
+    first.immigration.initialStock.newStream().nextLong() shouldBe second.immigration.initialStock.newStream().nextLong()
   }
